refactor(desktop): extract install action hook from mod button
